product

+++ database/migrations/2022_12_16_082353_create_product_options_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('product_options', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('product_id');
             $table->timestamps();
         });
     }
