+++ database/migrations/2023_02_08_145103_create_product_variants_table.php
@@ -15,10 +15,10 @@ return new class extends Migration
     {
         Schema::create('product_variants', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('product_id');
             $table->foreignId('sku_id')->constrained('product_skus')->cascadeOnDelete();
-            $table->unsignedInteger('product_option_id');
-            $table->unsignedInteger('product_option_value_id');
+            $table->unsignedBigInteger('product_option_id');
+            $table->unsignedBigInteger('product_option_value_id');
             $table->timestamps();
         });
     }
